fix: 修复门控MLP输出数据类型转换，使用输入dtype替代硬编码的float16
将输出张量的数据类型从硬编码的`tl.float16`改为动态使用输入块的数据类型`x_block.dtype`，提高kernel的灵活性和通用性，支持不同精度的输入数据。

+++ triton_ops/kernels/gated_mlp.py
@@ -143,7 +143,7 @@ def fused_gated_mlp_kernel(
     out_mask = (rows[:, None] < (batch_size * seq_len)) & (cols[None, :] < intermediate_dim)
 
     # Convert to output dtype
-    tl.store(out_ptrs, output.to(tl.float16), mask=out_mask)
+    tl.store(out_ptrs, output.to(x_block.dtype), mask=out_mask)
 
 
 def fused_gated_mlp(
